LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-expected.txt

Testing the TracksPanel class properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Basic properties
PASS tracksPanel.element.localName is "div"
PASS tracksPanel.element.className is "tracks-panel"
PASS tracksPanel.presented is false
PASS tracksPanel.rightX is 0

PASS successfullyParsed is true

TEST COMPLETE
